Commenting and reconstruction

+++ testing/instance/loadSingle.php
@@ -1,7 +1,27 @@
 <?php
 require_once '../../src/hicurl.php';
-unlink('history.txt');
-$hicurl=new Hicurl(['history'=>'history.txt']);
-$result=$hicurl->loadSingle('www.google.com');
-$hicurl->compileHistory();
-//echo $result['content'];
+if (!isset($_GET['getHistory'])) {
+	unlink('history.txt');
+	$hicurl=new Hicurl(['history'=>'history.txt']);
+	$result=$hicurl->loadSingle('www.google.com');
+	$hicurl->compileHistory();
+	?>
+		<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01//EN"
+		"http://www.w3.org/TR/html4/strict.dtd">
+		<html lang="en">
+		  <head>
+			<meta http-equiv="content-type" content="text/html; charset=utf-8">
+			<title>Hicurl Load Single Test</title>
+<!--			<link rel="stylesheet" type="text/css" href="style.css">-->
+			
+			<script src="https://code.jquery.com/jquery-2.1.3.min.js"></script><!--Hicurl requires jquery-->
+			<script src="../../src/hicurl.js"></script>
+			<script src="loadSingle.js"></script>
+		  </head>
+		  <body>
+		  </body>
+		</html>
+	<?php
+} else {
+	Hicurl::echoHistory('history.txt');
+}
